Split up integration test suite
Split up the integration test suite into symbolize, normalize, and
inspect modules. Doing so mirrors what we are already doing for
benchmarks and is just a more logical grouping all things considered.
